refactor(narrative/icon): Remove in favor of components/util/icon.
