+++ lib/components/admin/call-history-window.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { connect } from 'react-redux'
 
 import * as callTakerActions from '../../actions/call-taker'
-import Icon from '../narrative/icon'
+import Icon from '../util/icon'
 
 import CallRecord from './call-record'
 import DraggableWindow from './draggable-window'
