+++ lib/components/admin/call-record.js
@@ -2,8 +2,8 @@ import humanizeDuration from 'humanize-duration'
 import moment from 'moment'
 import React, { Component } from 'react'
 
-import Icon from '../narrative/icon'
 import {searchToQuery} from '../../util/call-taker'
+import Icon from '../util/icon'
 
 import CallTimeCounter from './call-time-counter'
 import QueryRecord from './query-record'
