+++ lib/components/admin/call-taker-controls.js
@@ -5,8 +5,8 @@ import * as apiActions from '../../actions/api'
 import * as callTakerActions from '../../actions/call-taker'
 import * as fieldTripActions from '../../actions/field-trip'
 import * as uiActions from '../../actions/ui'
-import Icon from '../narrative/icon'
 import { isModuleEnabled, Modules } from '../../util/config'
+import Icon from '../util/icon'
 
 import {
   CallHistoryButton,
