+++ lib/components/admin/draggable-window.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react'
 import Draggable from 'react-draggable'
 
-import Icon from '../narrative/icon'
+import Icon from '../util/icon'
 
 const noop = () => {}
