+++ lib/components/admin/field-trip-details.js
@@ -6,7 +6,6 @@ import { connect } from 'react-redux'
 import styled from 'styled-components'
 
 import * as fieldTripActions from '../../actions/field-trip'
-import Icon from '../narrative/icon'
 import {
   getActiveFieldTripRequest,
   getGroupSize,
@@ -14,6 +13,7 @@ import {
   PAYMENT_FIELDS,
   TICKET_TYPES
 } from '../../util/call-taker'
+import Icon from '../util/icon'
 
 import DraggableWindow from './draggable-window'
 import EditableSection from './editable-section'
