+++ lib/components/admin/field-trip-itinerary-group-size.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Badge } from 'react-bootstrap'
 
-import Icon from '../narrative/icon'
+import Icon from '../util/icon'
 
 export default function FieldTripGroupSize ({ itinerary }) {
   return itinerary.fieldTripGroupSize > 0 && (
