+++ lib/components/admin/field-trip-list.js
@@ -5,10 +5,10 @@ import { Badge, Button } from 'react-bootstrap'
 import { connect } from 'react-redux'
 
 import * as fieldTripActions from '../../actions/field-trip'
-import Icon from '../narrative/icon'
 import Loading from '../narrative/loading'
 import {getVisibleRequests, TABS} from '../../util/call-taker'
 import {FETCH_STATUS} from '../../util/constants'
+import Icon from '../util/icon'
 
 import {FieldTripRecordButton, WindowHeader} from './styled'
 import DraggableWindow from './draggable-window'
