+++ lib/components/admin/field-trip-notes.js
@@ -2,7 +2,8 @@ import React, { Component } from 'react'
 import { Badge, Button as BsButton } from 'react-bootstrap'
 import styled from 'styled-components'
 
-import Icon from '../narrative/icon'
+import Icon from '../util/icon'
+
 import {
   Button,
   Full,
