+++ lib/components/admin/mailables-window.js
@@ -3,8 +3,8 @@ import {Badge, Button} from 'react-bootstrap'
 import {connect} from 'react-redux'
 
 import * as callTakerActions from '../../actions/call-taker'
-import Icon from '../narrative/icon'
 import {getModuleConfig, isModuleEnabled, Modules} from '../../util/config'
+import Icon from '../util/icon'
 import {createLetter, LETTER_FIELDS} from '../../util/mailables'
 
 import {
