+++ lib/components/admin/styled.js
@@ -1,7 +1,7 @@
 import { Button as BsButton } from 'react-bootstrap'
 import styled, {css} from 'styled-components'
 
-import Icon from '../narrative/icon'
+import Icon from '../util/icon'
 
 import DefaultCounter from './call-time-counter'
