+++ lib/components/admin/trip-status.js
@@ -5,8 +5,8 @@ import { connect } from 'react-redux'
 
 import * as fieldTripActions from '../../actions/field-trip'
 import * as formActions from '../../actions/form'
-import Icon from '../narrative/icon'
 import { getTripFromRequest } from '../../util/call-taker'
+import Icon from '../util/icon'
 
 import {
   Bold,
