+++ lib/components/app/app-menu.js
@@ -5,11 +5,11 @@ import { connect } from 'react-redux'
 import { DropdownButton, MenuItem } from 'react-bootstrap'
 import { withRouter } from 'react-router'
 
-import Icon from '../narrative/icon'
 import * as callTakerActions from '../../actions/call-taker'
 import * as fieldTripActions from '../../actions/field-trip'
 import { MainPanelContent, setMainPanelContent } from '../../actions/ui'
 import { isModuleEnabled, Modules } from '../../util/config'
+import Icon from '../util/icon'
 
 // TODO: make menu items configurable via props/config
