+++ lib/components/form/batch-settings.js
@@ -5,7 +5,7 @@ import styled from 'styled-components'
 
 import * as apiActions from '../../actions/api'
 import * as formActions from '../../actions/form'
-import Icon from '../narrative/icon'
+import Icon from '../util/icon'
 import { hasValidLocation, getActiveSearch, getShowUserSettings } from '../../util/state'
 
 import BatchPreferences from './batch-preferences'
