+++ lib/components/form/mode-buttons.js
@@ -2,8 +2,8 @@ import React, { useContext } from 'react'
 import { OverlayTrigger, Tooltip } from 'react-bootstrap'
 import styled from 'styled-components'
 
-import Icon from '../narrative/icon'
 import { ComponentContext } from '../../util/contexts'
+import Icon from '../util/icon'
 
 import {buttonCss} from './batch-styled'
