+++ lib/components/form/user-settings.js
@@ -4,11 +4,11 @@ import React, { Component } from 'react'
 import { Button } from 'react-bootstrap'
 import { connect } from 'react-redux'
 
-import Icon from '../narrative/icon'
 import { forgetSearch, toggleTracking } from '../../actions/api'
 import { setQueryParam } from '../../actions/form'
 import { forgetPlace, forgetStop, setLocation } from '../../actions/map'
 import { setViewedStop } from '../../actions/ui'
+import Icon from '../util/icon'
 
 const { formatStoredPlaceName, getDetailText, matchLatLon } = coreUtils.map
 const { summarizeQuery } = coreUtils.query
