+++ lib/components/form/user-trip-settings.js
@@ -3,12 +3,12 @@ import React, { Component } from 'react'
 import { Button } from 'react-bootstrap'
 import { connect } from 'react-redux'
 
-import Icon from '../narrative/icon'
 import {
   clearDefaultSettings,
   resetForm,
   storeDefaultSettings
 } from '../../actions/form'
+import Icon from '../util/icon'
 
 /**
  * This component contains the `Remember/Forget my trip options` and `Restore defaults` commands
