+++ lib/components/mobile/batch-results-screen.js
@@ -6,8 +6,8 @@ import styled, { css } from 'styled-components'
 
 import * as uiActions from '../../actions/ui'
 import Map from '../map/map'
-import Icon from '../narrative/icon'
 import NarrativeItineraries from '../narrative/narrative-itineraries'
+import Icon from '../util/icon'
 import {
   getActiveItineraries,
   getActiveSearch,
