+++ lib/components/mobile/navigation-bar.js
@@ -6,9 +6,9 @@ import { connect } from 'react-redux'
 import { setMobileScreen } from '../../actions/ui'
 import AppMenu from '../app/app-menu'
 import NavLoginButtonAuth0 from '../../components/user/nav-login-button-auth0'
-import Icon from '../narrative/icon'
 import { accountLinks, getAuth0Config } from '../../util/auth'
 import { ComponentContext } from '../../util/contexts'
+import Icon from '../util/icon'
 
 class MobileNavigationBar extends Component {
   static propTypes = {
