+++ lib/components/narrative/default/access-leg.js
@@ -3,7 +3,7 @@ import { humanizeDistanceString } from '@opentripplanner/humanize-distance'
 import PropTypes from 'prop-types'
 import React, {Component} from 'react'
 
-import Icon from '../icon'
+import Icon from '../../util/icon'
 import LegDiagramPreview from '../leg-diagram-preview'
 
 /**
