+++ lib/components/narrative/default/transit-leg.js
@@ -2,7 +2,7 @@ import coreUtils from '@opentripplanner/core-utils'
 import PropTypes from 'prop-types'
 import React, { Component } from 'react'
 
-import Icon from '../icon'
+import Icon from '../../util/icon'
 import ViewTripButton from '../../viewers/view-trip-button'
 import ViewStopButton from '../../viewers/view-stop-button'
