+++ lib/components/narrative/itinerary-carousel.js
@@ -8,8 +8,8 @@ import SwipeableViews from 'react-swipeable-views'
 import { setActiveItinerary, setActiveLeg, setActiveStep } from '../../actions/narrative'
 import { ComponentContext } from '../../util/contexts'
 import { getActiveItineraries, getActiveSearch } from '../../util/state'
+import Icon from '../util/icon'
 
-import Icon from './icon'
 import Loading from './loading'
 
 class ItineraryCarousel extends Component {
