+++ lib/components/narrative/loading.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react'
 
-import Icon from './icon'
+import Icon from '../util/icon'
 
 export default class Loading extends Component {
   render () {
