+++ lib/components/narrative/mode-icon.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 
-import Icon from './icon'
+import Icon from '../util/icon'
 
 export default class ModeIcon extends Component {
   static propTypes = {
