+++ lib/components/narrative/narrative-itineraries-errors.js
@@ -1,7 +1,7 @@
 import { getCompanyIcon } from '@opentripplanner/icons/lib/companies'
 import styled from 'styled-components'
 
-import Icon from '../narrative/icon'
+import Icon from '../util/icon'
 import { getErrorMessage } from '../../util/state'
 
 const IssueContainer = styled.div`
