+++ lib/components/narrative/narrative-itineraries-header.js
@@ -1,7 +1,7 @@
 import styled from 'styled-components'
 import { FormattedMessage, useIntl } from 'react-intl'
 
-import Icon from '../narrative/icon'
+import Icon from '../util/icon'
 
 import PlanFirstLastButtons from './plan-first-last-buttons'
 import SaveTripButton from './save-trip-button'
