+++ lib/components/user/monitored-trip/trip-notifications-pane.js
@@ -3,7 +3,7 @@ import React, { Component } from 'react'
 import { Alert, FormControl, Glyphicon } from 'react-bootstrap'
 import styled from 'styled-components'
 
-import Icon from '../../narrative/icon'
+import Icon from '../../util/icon'
 
 const notificationChannelLabels = {
   email: 'email',
