+++ lib/components/user/places/favorite-place-row.js
@@ -4,7 +4,7 @@ import { Button } from 'react-bootstrap'
 import styled, { css } from 'styled-components'
 
 import { LinkContainerWithQuery } from '../../form/connected-links'
-import Icon from '../../narrative/icon'
+import Icon from '../../util/icon'
 
 const FIELD_HEIGHT_PX = '60px'
