+++ lib/components/user/places/place-editor.js
@@ -10,7 +10,7 @@ import {
 } from 'react-bootstrap'
 import styled from 'styled-components'
 
-import Icon from '../../narrative/icon'
+import Icon from '../../util/icon'
 import { getErrorStates } from '../../../util/ui'
 import { CUSTOM_PLACE_TYPES, isHomeOrWork } from '../../../util/user'
