+++ lib/components/user/styled.js
@@ -1,7 +1,7 @@
 import { Panel } from 'react-bootstrap'
 import styled from 'styled-components'
 
-import Icon from '../narrative/icon'
+import Icon from '../util/icon'
 
 export const PageHeading = styled.h2`
   margin: 10px 0px 45px 0px;
