+++ lib/components/util/icon.js
@@ -22,8 +22,9 @@ const StyledFontAwesome = styled(FontAwesome)`
  * and that should work for both left-to-right and right-to-left layouts.
  * Other props from FontAwesome are passed to that component.
  */
-const Icon = ({ type, withSpace, ...props }) => (
+const Icon = ({ fixedWidth = true, type, withSpace, ...props }) => (
   <StyledFontAwesome
+    fixedWidth={fixedWidth}
     name={type}
     withSpace={withSpace}
     {...props}
