+++ lib/components/viewers/live-stop-times.js
@@ -3,7 +3,7 @@ import 'moment-timezone'
 import coreUtils from '@opentripplanner/core-utils'
 import React, { Component } from 'react'
 
-import Icon from '../narrative/icon'
+import Icon from '../util/icon'
 import {
   getRouteIdForPattern,
   getStopTimesByPattern,
