+++ lib/components/viewers/pattern-row.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react'
 import { VelocityTransitionGroup } from 'velocity-react'
 
-import Icon from '../narrative/icon'
+import Icon from '../util/icon'
 import { stopTimeComparator } from '../../util/viewer'
 
 import RealtimeStatusLabel from './realtime-status-label'
