+++ lib/components/viewers/route-viewer.js
@@ -6,7 +6,7 @@ import { VelocityTransitionGroup } from 'velocity-react'
 import { connect } from 'react-redux'
 import styled from 'styled-components'
 
-import Icon from '../narrative/icon'
+import Icon from '../util/icon'
 import { setMainPanelContent, setViewedRoute } from '../../actions/ui'
 import { findRoutes, findRoute } from '../../actions/api'
 import { ComponentContext } from '../../util/contexts'
