+++ lib/components/viewers/stop-time-cell.js
@@ -4,7 +4,7 @@ import coreUtils from '@opentripplanner/core-utils'
 import PropTypes from 'prop-types'
 import React from 'react'
 
-import Icon from '../narrative/icon'
+import Icon from '../util/icon'
 import { getSecondsUntilDeparture } from '../../util/viewer'
 
 const { formatDuration, formatSecondsAfterMidnight } = coreUtils.time
