+++ lib/components/viewers/stop-viewer.js
@@ -11,7 +11,7 @@ import styled from 'styled-components'
 import * as apiActions from '../../actions/api'
 import * as mapActions from '../../actions/map'
 import * as uiActions from '../../actions/ui'
-import Icon from '../narrative/icon'
+import Icon from '../util/icon'
 import { getShowUserSettings, getStopViewerConfig } from '../../util/state'
 
 import LiveStopTimes from './live-stop-times'
