+++ lib/components/viewers/trip-viewer.js
@@ -6,7 +6,7 @@ import React, { Component } from 'react'
 import { Button, Label } from 'react-bootstrap'
 import { connect } from 'react-redux'
 
-import Icon from '../narrative/icon'
+import Icon from '../util/icon'
 import { setViewedTrip } from '../../actions/ui'
 import { findTrip } from '../../actions/api'
 import { setLocation } from '../../actions/map'
